testbench for mod

+++ src/testmod.h
@@ -1,7 +1,7 @@
 #ifndef __TESTMOD_H__
 #define __TESTMOD_H__
 static void set_mod_reg(int val);
-static int add_mod(int val1,int val2);
-static int sub_mod(int val1,int val2);
-static int mul_mod(int val1,int val2);
+static long long add_mod(long long val1,long long val2);
+static long long sub_mod(long long val1,long long val2);
+static long long mul_mod(long long val1,long long val2);
 #endif /* __TESTMOD_H__ */
